Basic camping functionality

+++ trunk/source/ZoneServer/Spawns/SpawnStructs.h
@@ -563,7 +563,7 @@ struct SpawnInfoStruct {
 
 	SpawnInfoStruct() {
 		//Hack to zero out most of the data allowing default values for the eq2 colors
-		memset(&model_type, 0, reinterpret_cast<uint64_t>(equipment_colors) - reinterpret_cast<uint64_t>(&model_type));
+		memset(&model_type, 0, reinterpret_cast<size_t>(equipment_colors) - reinterpret_cast<size_t>(&model_type));
 	}
 
 protected:
